requirements.txt
Flask
Flask-Cors
mysql-connector-python
Flask-SQLAlchemy
Flask-Migrate